Store the number of hydro and gravity particles as a size_t.

+++ src/velociraptor_interface.c
@@ -151,8 +151,8 @@ void velociraptor_invoke(struct engine *e) {
 
     struct space *s = e->s;
     struct gpart *gparts = s->gparts;
-    const int nr_gparts = s->nr_gparts;
-    const int nr_hydro_parts = s->nr_parts;
+    const size_t nr_gparts = s->nr_gparts;
+    const size_t nr_hydro_parts = s->nr_parts;
     const int nr_cells = s->nr_cells;
     int *cell_node_ids;
